requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
APScheduler==4.0.0a5
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
pip==24.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.1
serial==0.0.97
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity>=8.0,<10.0
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
